pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "django-sslcommerz"
dynamic = ["version"]
description = "Django integration package for SSLCOMMERZ Payment Gateway API v4"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["django", "sslcommerz", "payment", "gateway", "bangladesh"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Point-Of-Sale",
]
requires-python = ">=3.8"
dependencies = [
    "Django>=3.2,<5.0",
    "requests>=2.25.0",
    "python-decouple>=3.4",
]

[project.optional-dependencies]
drf = ["djangorestframework>=3.12.0"]
dev = [
    "pytest>=6.0.0",
    "pytest-django>=4.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "isort>=5.0.0",
    "mypy>=0.910",
    "django-stubs>=1.9.0",
    "bandit>=1.7.0",
    "safety>=1.10.0",
    "pre-commit>=2.15.0",
]

[project.urls]
Homepage = "https://github.com/your-username/django-sslcommerz"
Documentation = "https://django-sslcommerz.readthedocs.io/"
Repository = "https://github.com/your-username/django-sslcommerz.git"
"Bug Reports" = "https://github.com/your-username/django-sslcommerz/issues"
Changelog = "https://github.com/your-username/django-sslcommerz/blob/main/CHANGELOG.md"

[tool.setuptools_scm]
write_to = "sslcommerz/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = "sslcommerz"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "test_project.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_project.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=sslcommerz",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=85",
]

[tool.coverage.run]
source = ["sslcommerz"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*Protocol.*:",
]

[tool.bandit]
exclude_dirs = ["tests", "test_project"]
skips = ["B101", "B601"]
